Incident-Response/Tools/cyphon/requirements.txt
amqp==2.3.2
anyjson==0.3.3
backoff==1.6.0
beautifulsoup4==4.6.1
billiard==3.5.0.4
bleach==2.1.3
boto3==1.9.9
celery==4.2.1
django-autocomplete-light==2.3.3  # pyup: >=2.0.0,<3.0.0
django-ckeditor==5.6.1
django-constance==2.3.1
django-extensions==2.1.0
django-filter==1.1.0
django-grappelli==2.11.1
django-mailbox==4.7.1
django-picklefield==1.0.0
django-s3-storage==0.12.4
Django==1.11.15  # pyup: >=1.0.0,<2.0.0
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
docutils==0.14
drfdocs==0.0.11
ec2-metadata==1.6.0
elasticsearch==5.5.3  # pyup: >=5.0.0,<6.0.0
feedparser==5.2.1
geoip2==2.8.0
geopy==1.17.0
gunicorn==19.8.1
html5lib==1.0.1
httmock==1.2.6
httplib2==0.11.3
jira==2.0.0
kombu==4.2.1
langdetect==1.0.7
maxminddb==1.4.1
nltk==3.3
oauth2==1.9.0.post1
oauthlib==2.1.0
pika==0.12.0
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pymongo==3.7.1
python-dateutil==2.7.3
pytz==2018.5
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
requests==2.20.0
sauceclient==1.0.0
selenium==3.13.0
six==1.11.0
testfixtures==6.2.0
textblob==0.15.1
tlslite==0.4.9
tweepy==3.6.0
urllib3==1.23
pipdeptree==0.13.0
